rename them back to the correct name to reflect uppercase changes
