+++ src/controllers/ItemController.ts
@@ -1,4 +1,4 @@
-import { Items } from '../entity/TempItems'
+import { Items } from '../entity/Items'
 import { ItemRepository } from '../models/repository/ItemRepository'
 import { LocationRepository } from '../models/repository/LocationRepository'
