+++ src/entity/Locations.ts
